Added access to a building to house Lon Jatham.
I added a new zone to the ados.xml file which has been modelled after
the library zone. Portals were added to the large building in Ados city
with two doors, so that it could be used in the game. Finally,I removed
the collisions from the doors in the city.tmx file so that the player
could walk up to the door.