Added Lon Jatham NPC in UoM Campus.
The NPC has the same looks as the librarian, or is it the other way
around? Jatham should answer the questions the player has to decide
whether The University of Manchester is a good fit for them. Lon should
also be able to provide any programmer a tip to help them in their
journey.