All JUnit Tests now pass
A JUnit test, which makes sure that pairs of portal references match so
that they can be used to go back and forth, was failing due to a portal
reference for house 69 which was altered in ados.xml not being changed
back to its original state when it was determined we would not be using
that building. The issue has been resolved by changing the incorrect
reference back to its original state so that it matched the other one in
the pair.