Edited comments to increase code legibility.
The comments used to help identify the parts of ados.xml were removed,
so as to not confuse persons that may later work on the code and perhaps
use their own place holders/markers. A comment was added to the
LonJathamNPC code to explain why there is a reference to the Ados
library and an author named Hendrik.