Fixed Wofol Scroll test
The test failed due to errors. The test now initialises the whole of
semos. It checks whether the user has been to wofol and that they can't
use the scroll until they have. It then moves the player to wofol and
back and then checks that the player has now been transported there.