Created handler files for XML
Code has to be finished in DefaultAction.java. I dind't add the optional
field handling, therefore that has to be done. XMLSlashAction.java so
far is just a copy of Invisible action. I added a constructor to it,
just to get rid of errors. The loader file also doesn't deal with the
optional fields, therefore that also has to be done.