Migrated most functions.
All functions but the ones that work with 'remainder' elements have been
migrated. If they just had a type and put elements into the message to
the server then they were treated as 'simple' cases and handled by the
default XMLSlashAction. On the other hand, if they had conditions for
the parameters they are handled by their own class, which is always a
subclass of XMLSlashAction.